# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sizes:
          type: array
          items:
            type: object
            properties:
              size:
                type: string
              quantity:
                type: integer
            required:
              - size
              - quantity
        inventory:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
              properties:
                hex:
                  type: string
                qty:
                  type: integer
              required:
                - hex
                - qty
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - stock
        - sizes
        - createdAt
        - updatedAt
...
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        mrp:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        categoryId:
          type: integer
          nullable: true
        rating:
          type: string
        reviewCount:
          type: integer
        reviewText:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: object
            properties:
              size:
                type: string
              quantity:
                type: integer
            required:
              - size
              - quantity
        inventory:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
              properties:
                hex:
                  type: string
                qty:
                  type: integer
              required:
                - hex
                - qty
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...